+++ pilot/pkg/networking/core/v1alpha3/cluster_builder.go
@@ -20,31 +20,51 @@ import (
 	cluster "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	endpoint "github.com/envoyproxy/go-control-plane/envoy/config/endpoint/v3"
+	auth "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
+	http "github.com/envoyproxy/go-control-plane/envoy/extensions/upstreams/http/v3"
 	"github.com/gogo/protobuf/types"
+	"github.com/golang/protobuf/ptypes"
+	"github.com/golang/protobuf/ptypes/any"
 	"github.com/golang/protobuf/ptypes/duration"
+	structpb "github.com/golang/protobuf/ptypes/struct"
 	"github.com/golang/protobuf/ptypes/wrappers"
 
+	meshconfig "istio.io/api/mesh/v1alpha1"
 	networking "istio.io/api/networking/v1alpha3"
+	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/util"
+	authn_model "istio.io/istio/pilot/pkg/security/model"
+	"istio.io/istio/pilot/pkg/util/sets"
+	v3 "istio.io/istio/pilot/pkg/xds/v3"
 	"istio.io/istio/pkg/config"
 	"istio.io/istio/pkg/config/labels"
 	"istio.io/istio/pkg/util/gogo"
+	"istio.io/pkg/log"
 )
 
 var defaultDestinationRule = networking.DestinationRule{}
+
+var istioMtlsTransportSocketMatch = &structpb.Struct{
+	Fields: map[string]*structpb.Value{
+		model.TLSModeLabelShortname: {Kind: &structpb.Value_StringValue{StringValue: model.IstioMutualTLSModeLabel}},
+	},
+}
 
 // ClusterBuilder interface provides an abstraction for building Envoy Clusters.
 type ClusterBuilder struct {
 	proxy *model.Proxy
 	push  *model.PushContext
+	// httpProtocolOptions stores the HttpProtocolOptions per cluster.
+	httpProtocolOptions map[string]*http.HttpProtocolOptions
 }
 
 // NewClusterBuilder builds an instance of ClusterBuilder.
 func NewClusterBuilder(proxy *model.Proxy, push *model.PushContext) *ClusterBuilder {
 	return &ClusterBuilder{
-		proxy: proxy,
-		push:  push,
+		proxy:               proxy,
+		push:                push,
+		httpProtocolOptions: map[string]*http.HttpProtocolOptions{},
 	}
 }
 
@@ -76,7 +96,7 @@ func (cb *ClusterBuilder) applyDestinationRule(c *cluster.Cluster, clusterMode C
 	// merge with applicable port level traffic policy settings
 	opts.policy = MergeTrafficPolicy(nil, opts.policy, opts.port)
 	// Apply traffic policy for the main default cluster.
-	applyTrafficPolicy(opts)
+	cb.applyTrafficPolicy(opts)
 
 	// Apply EdsConfig if needed. This should be called after traffic policy is applied because, traffic policy might change
 	// discovery type.
@@ -124,7 +144,7 @@ func (cb *ClusterBuilder) applyDestinationRule(c *cluster.Cluster, clusterMode C
 		if len(cb.push.Mesh.OutboundClusterStatName) != 0 {
 			subsetCluster.AltStatName = util.BuildStatPrefix(cb.push.Mesh.OutboundClusterStatName, string(service.Hostname), subset.Name, port, service.Attributes)
 		}
-		setUpstreamProtocol(cb.proxy, subsetCluster, port, model.TrafficDirectionOutbound)
+		cb.setUpstreamProtocol(cb.proxy, subsetCluster, port, model.TrafficDirectionOutbound)
 
 		// Apply traffic policy for subset cluster with the destination rule traffic policy.
 		opts.cluster = subsetCluster
@@ -133,7 +153,7 @@ func (cb *ClusterBuilder) applyDestinationRule(c *cluster.Cluster, clusterMode C
 		// If subset has a traffic policy, apply it so that it overrides the destination rule traffic policy.
 		opts.policy = MergeTrafficPolicy(destinationRule.TrafficPolicy, subset.TrafficPolicy, opts.port)
 		// Apply traffic policy for the subset cluster.
-		applyTrafficPolicy(opts)
+		cb.applyTrafficPolicy(opts)
 
 		maybeApplyEdsConfig(subsetCluster)
 
@@ -144,6 +164,391 @@ func (cb *ClusterBuilder) applyDestinationRule(c *cluster.Cluster, clusterMode C
 		subsetClusters = append(subsetClusters, subsetCluster)
 	}
 	return subsetClusters
+}
+
+func (cb *ClusterBuilder) applyTrafficPolicy(opts buildClusterOpts) {
+	connectionPool, outlierDetection, loadBalancer, tls := selectTrafficPolicyComponents(opts.policy)
+	if opts.direction == model.TrafficDirectionOutbound && connectionPool != nil && connectionPool.Http != nil && connectionPool.Http.UseClientProtocol {
+		// Use downstream protocol. If the incoming traffic use HTTP 1.1, the
+		// upstream cluster will use HTTP 1.1, if incoming traffic use HTTP2,
+		// the upstream cluster will use HTTP2.
+		cb.setUseDownstreamProtocol(opts.cluster)
+	}
+	// Connection pool settings are applicable for both inbound and outbound clusters.
+	cb.applyConnectionPool(opts.mesh, opts.cluster, connectionPool)
+	if opts.direction != model.TrafficDirectionInbound {
+		cb.applyH2Upgrade(opts, connectionPool)
+		applyOutlierDetection(opts.cluster, outlierDetection)
+		applyLoadBalancer(opts.cluster, loadBalancer, opts.port, opts.proxy, opts.mesh)
+	}
+	if opts.cluster.GetType() == cluster.Cluster_ORIGINAL_DST {
+		opts.cluster.LbPolicy = cluster.Cluster_CLUSTER_PROVIDED
+	}
+
+	if opts.clusterMode != SniDnatClusterMode && opts.direction != model.TrafficDirectionInbound {
+		autoMTLSEnabled := opts.mesh.GetEnableAutoMtls().Value
+		var mtlsCtxType mtlsContextType
+		tls, mtlsCtxType = buildAutoMtlsSettings(tls, opts.serviceAccounts, opts.istioMtlsSni, opts.proxy,
+			autoMTLSEnabled, opts.meshExternal, opts.serviceMTLSMode)
+		cb.applyUpstreamTLSSettings(&opts, tls, mtlsCtxType)
+	}
+}
+
+// applyH2Upgrade function will upgrade outbound cluster to http2 if specified by configuration.
+func (cb *ClusterBuilder) applyH2Upgrade(opts buildClusterOpts, connectionPool *networking.ConnectionPoolSettings) {
+	if cb.shouldH2Upgrade(opts.cluster.Name, opts.direction, opts.port, opts.mesh, connectionPool) {
+		cb.setH2Options(opts.cluster)
+	}
+}
+
+// shouldH2Upgrade function returns true if the cluster  should be upgraded to http2.
+func (cb *ClusterBuilder) shouldH2Upgrade(clusterName string, direction model.TrafficDirection, port *model.Port, mesh *meshconfig.MeshConfig,
+	connectionPool *networking.ConnectionPoolSettings) bool {
+	if direction != model.TrafficDirectionOutbound {
+		return false
+	}
+
+	// TODO (mjog)
+	// Upgrade if tls.GetMode() == networking.TLSSettings_ISTIO_MUTUAL
+	override := networking.ConnectionPoolSettings_HTTPSettings_DEFAULT
+	if connectionPool != nil && connectionPool.Http != nil {
+		override = connectionPool.Http.H2UpgradePolicy
+	}
+	// If user wants an upgrade at destination rule/port level that means he is sure that
+	// it is a Http port - upgrade in such case. This is useful incase protocol sniffing is
+	// enabled and user wants to upgrade/preserve http protocol from client.
+	if override == networking.ConnectionPoolSettings_HTTPSettings_UPGRADE {
+		log.Debugf("Upgrading cluster: %v (%v %v)", clusterName, mesh.H2UpgradePolicy, override)
+		return true
+	}
+
+	// Do not upgrade non-http ports
+	// This also ensures that we are only upgrading named ports so that
+	// EnableProtocolSniffingForInbound does not interfere.
+	// protocol sniffing uses Cluster_USE_DOWNSTREAM_PROTOCOL.
+	// Therefore if the client upgrades connection to http2, the server will send h2 stream to the application,
+	// even though the application only supports http 1.1.
+	if port != nil && !port.Protocol.IsHTTP() {
+		return false
+	}
+
+	if !h2UpgradeMap[upgradeTuple{mesh.H2UpgradePolicy, override}] {
+		log.Debugf("Not upgrading cluster: %v (%v %v)", clusterName, mesh.H2UpgradePolicy, override)
+		return false
+	}
+
+	log.Debugf("Upgrading cluster: %v (%v %v)", clusterName, mesh.H2UpgradePolicy, override)
+	return true
+}
+
+// setH2Options make the cluster an h2 cluster by setting http2ProtocolOptions.
+func (cb *ClusterBuilder) setH2Options(cluster *cluster.Cluster) {
+	if cluster == nil {
+		return
+	}
+	if cb.httpProtocolOptions[cluster.Name] == nil {
+		cb.httpProtocolOptions[cluster.Name] = &http.HttpProtocolOptions{}
+	}
+	options := cb.httpProtocolOptions[cluster.Name]
+	if options.UpstreamHttpProtocolOptions == nil {
+		options.UpstreamProtocolOptions = &http.HttpProtocolOptions_ExplicitHttpConfig_{
+			ExplicitHttpConfig: &http.HttpProtocolOptions_ExplicitHttpConfig{
+				ProtocolConfig: &http.HttpProtocolOptions_ExplicitHttpConfig_Http2ProtocolOptions{
+					Http2ProtocolOptions: http2ProtocolOptions(),
+				},
+			},
+		}
+	}
+}
+
+func (cb *ClusterBuilder) setUseDownstreamProtocol(cluster *cluster.Cluster) {
+	if cb.httpProtocolOptions[cluster.Name] == nil {
+		cb.httpProtocolOptions[cluster.Name] = &http.HttpProtocolOptions{}
+	}
+	options := cb.httpProtocolOptions[cluster.Name]
+	options.UpstreamProtocolOptions = &http.HttpProtocolOptions_UseDownstreamProtocolConfig{
+		UseDownstreamProtocolConfig: &http.HttpProtocolOptions_UseDownstreamHttpConfig{
+			Http2ProtocolOptions: http2ProtocolOptions(),
+		},
+	}
+}
+
+func (cb *ClusterBuilder) applyUpstreamTLSSettings(opts *buildClusterOpts, tls *networking.ClientTLSSettings, mtlsCtxType mtlsContextType) {
+	if tls == nil {
+		return
+	}
+
+	c := opts.cluster
+
+	tlsContext, err := cb.buildUpstreamClusterTLSContext(opts, tls)
+	if err != nil {
+		log.Errorf("failed to build Upstream TLSContext: %s", err.Error())
+		return
+	}
+
+	if tlsContext != nil {
+		c.TransportSocket = &core.TransportSocket{
+			Name:       util.EnvoyTLSSocketName,
+			ConfigType: &core.TransportSocket_TypedConfig{TypedConfig: util.MessageToAny(tlsContext)},
+		}
+	}
+
+	// For headless service, discover type will be `Cluster_ORIGINAL_DST`
+	// Apply auto mtls to clusters excluding these kind of headless service
+	if c.GetType() != cluster.Cluster_ORIGINAL_DST {
+		// convert to transport socket matcher if the mode was auto detected
+		if tls.Mode == networking.ClientTLSSettings_ISTIO_MUTUAL && mtlsCtxType == autoDetected {
+			transportSocket := c.TransportSocket
+			c.TransportSocket = nil
+			c.TransportSocketMatches = []*cluster.Cluster_TransportSocketMatch{
+				{
+					Name:            "tlsMode-" + model.IstioMutualTLSModeLabel,
+					Match:           istioMtlsTransportSocketMatch,
+					TransportSocket: transportSocket,
+				},
+				defaultTransportSocketMatch,
+			}
+		}
+	}
+}
+
+func (cb *ClusterBuilder) buildUpstreamClusterTLSContext(opts *buildClusterOpts, tls *networking.ClientTLSSettings) (*auth.UpstreamTlsContext, error) {
+	c := opts.cluster
+	proxy := opts.proxy
+
+	// Hack to avoid egress sds cluster config generation for sidecar when
+	// CredentialName is set in DestinationRule
+	if tls.CredentialName != "" && proxy.Type == model.SidecarProxy {
+		if tls.Mode == networking.ClientTLSSettings_SIMPLE || tls.Mode == networking.ClientTLSSettings_MUTUAL {
+			return nil, nil
+		}
+	}
+
+	var tlsContext *auth.UpstreamTlsContext
+
+	switch tls.Mode {
+	case networking.ClientTLSSettings_DISABLE:
+		tlsContext = nil
+	case networking.ClientTLSSettings_ISTIO_MUTUAL:
+
+		tlsContext = &auth.UpstreamTlsContext{
+			CommonTlsContext: &auth.CommonTlsContext{},
+			Sni:              tls.Sni,
+		}
+
+		// For ISTIO_MUTUAL_TLS we serve certificates from two well known locations.
+		// - Proxy Metadata: These are certs being mounted from within the pod. Rather
+		//   than reading directly in Envoy, which does not support rotation, we will
+		//   serve them over SDS by reading the files.
+		// - Default SDS Locations served with resource names "default" and "ROOTCA".
+
+		// buildIstioMutualTLS would have populated the TLSSettings with proxy metadata.
+		// Use it if it exists otherwise fallback to default SDS locations.
+		metadataSDS := model.SdsCertificateConfig{
+			CertificatePath:   tls.ClientCertificate,
+			PrivateKeyPath:    tls.PrivateKey,
+			CaCertificatePath: tls.CaCertificates,
+		}
+		metadataCerts := metadataSDS.IsRootCertificate() && metadataSDS.IsKeyCertificate()
+
+		tlsContext.CommonTlsContext.TlsCertificateSdsSecretConfigs = append(tlsContext.CommonTlsContext.TlsCertificateSdsSecretConfigs,
+			authn_model.ConstructSdsSecretConfig(model.GetOrDefault(metadataSDS.GetResourceName(), authn_model.SDSDefaultResourceName), proxy))
+
+		tlsContext.CommonTlsContext.ValidationContextType = &auth.CommonTlsContext_CombinedValidationContext{
+			CombinedValidationContext: &auth.CommonTlsContext_CombinedCertificateValidationContext{
+				DefaultValidationContext: &auth.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch(tls.SubjectAltNames)},
+				ValidationContextSdsSecretConfig: authn_model.ConstructSdsSecretConfig(model.GetOrDefault(metadataSDS.GetRootResourceName(),
+					authn_model.SDSRootResourceName), proxy),
+			},
+		}
+		// Set default SNI of cluster name for istio_mutual if sni is not set.
+		if len(tls.Sni) == 0 {
+			tlsContext.Sni = c.Name
+		}
+
+		// `istio-peer-exchange` alpn is only used when using mtls communication between peers.
+		// We add `istio-peer-exchange` to the list of alpn strings.
+		// The code has repeated snippets because We want to use predefined alpn strings for efficiency.
+		switch {
+		case metadataCerts:
+			if cb.httpProtocolOptions[c.Name] != nil && cb.httpProtocolOptions[c.Name].UpstreamProtocolOptions != nil {
+				// This is HTTP/2 cluster, advertise it with ALPN.
+				tlsContext.CommonTlsContext.AlpnProtocols = util.ALPNH2Only
+			}
+		default:
+			if cb.httpProtocolOptions[c.Name] != nil && cb.httpProtocolOptions[c.Name].UpstreamProtocolOptions != nil {
+				// This is HTTP/2 in-mesh cluster, advertise it with ALPN.
+				tlsContext.CommonTlsContext.AlpnProtocols = util.ALPNInMeshH2WithMxc
+			} else {
+				// This is in-mesh cluster, advertise it with ALPN.
+				tlsContext.CommonTlsContext.AlpnProtocols = util.ALPNInMeshWithMxc
+			}
+		}
+	case networking.ClientTLSSettings_SIMPLE:
+		tlsContext = &auth.UpstreamTlsContext{
+			CommonTlsContext: &auth.CommonTlsContext{},
+			Sni:              tls.Sni,
+		}
+
+		if tls.CredentialName != "" {
+			tlsContext = &auth.UpstreamTlsContext{
+				CommonTlsContext: &auth.CommonTlsContext{},
+				Sni:              tls.Sni,
+			}
+			// If  credential name is specified at Destination Rule config and originating node is egress gateway, create
+			// SDS config for egress gateway to fetch key/cert at gateway agent.
+			authn_model.ApplyCustomSDSToClientCommonTLSContext(tlsContext.CommonTlsContext, tls)
+		} else {
+			// If CredentialName is not set fallback to files specified in DR.
+			res := model.SdsCertificateConfig{
+				CaCertificatePath: model.GetOrDefault(proxy.Metadata.TLSClientRootCert, tls.CaCertificates),
+			}
+
+			// If tls.CaCertificate or CaCertificate in Metadata isn't configured don't set up SdsSecretConfig
+			if !res.IsRootCertificate() {
+				tlsContext.CommonTlsContext.ValidationContextType = &auth.CommonTlsContext_ValidationContext{}
+			} else {
+				tlsContext.CommonTlsContext.ValidationContextType = &auth.CommonTlsContext_CombinedValidationContext{
+					CombinedValidationContext: &auth.CommonTlsContext_CombinedCertificateValidationContext{
+						DefaultValidationContext:         &auth.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch(tls.SubjectAltNames)},
+						ValidationContextSdsSecretConfig: authn_model.ConstructSdsSecretConfig(res.GetRootResourceName(), proxy),
+					},
+				}
+			}
+		}
+
+		if cb.httpProtocolOptions[c.Name] != nil && cb.httpProtocolOptions[c.Name].UpstreamProtocolOptions != nil {
+			// This is HTTP/2 cluster, advertise it with ALPN.
+			tlsContext.CommonTlsContext.AlpnProtocols = util.ALPNH2Only
+		}
+
+	case networking.ClientTLSSettings_MUTUAL:
+		tlsContext = &auth.UpstreamTlsContext{
+			CommonTlsContext: &auth.CommonTlsContext{},
+			Sni:              tls.Sni,
+		}
+		if tls.CredentialName != "" {
+			// If  credential name is specified at Destination Rule config and originating node is egress gateway, create
+			// SDS config for egress gateway to fetch key/cert at gateway agent.
+			authn_model.ApplyCustomSDSToClientCommonTLSContext(tlsContext.CommonTlsContext, tls)
+		} else {
+			// If CredentialName is not set fallback to file based approach
+			if tls.ClientCertificate == "" || tls.PrivateKey == "" {
+				err := fmt.Errorf("failed to apply tls setting for %s: client certificate and private key must not be empty",
+					c.Name)
+				return nil, err
+			}
+
+			var res model.SdsCertificateConfig
+			if features.AllowMetadataCertsInMutualTLS {
+				// These are certs being mounted from within the pod and specified in Metadata.
+				// Rather than reading directly in Envoy, which does not support rotation, we will
+				// serve them over SDS by reading the files. This is only enabled for temporary migration.
+				res = model.SdsCertificateConfig{
+					CertificatePath:   model.GetOrDefault(proxy.Metadata.TLSClientCertChain, tls.ClientCertificate),
+					PrivateKeyPath:    model.GetOrDefault(proxy.Metadata.TLSClientKey, tls.PrivateKey),
+					CaCertificatePath: model.GetOrDefault(proxy.Metadata.TLSClientRootCert, tls.CaCertificates),
+				}
+			} else {
+				// These are certs being mounted from within the pod and specified in Destination Rules.
+				// Rather than reading directly in Envoy, which does not support rotation, we will
+				// serve them over SDS by reading the files.
+				res = model.SdsCertificateConfig{
+					CertificatePath:   tls.ClientCertificate,
+					PrivateKeyPath:    tls.PrivateKey,
+					CaCertificatePath: tls.CaCertificates,
+				}
+			}
+			tlsContext.CommonTlsContext.TlsCertificateSdsSecretConfigs = append(tlsContext.CommonTlsContext.TlsCertificateSdsSecretConfigs,
+				authn_model.ConstructSdsSecretConfig(res.GetResourceName(), proxy))
+
+			// If tls.CaCertificate or CaCertificate in Metadata isn't configured don't set up RootSdsSecretConfig
+			if !res.IsRootCertificate() {
+				tlsContext.CommonTlsContext.ValidationContextType = &auth.CommonTlsContext_ValidationContext{}
+			} else {
+				tlsContext.CommonTlsContext.ValidationContextType = &auth.CommonTlsContext_CombinedValidationContext{
+					CombinedValidationContext: &auth.CommonTlsContext_CombinedCertificateValidationContext{
+						DefaultValidationContext:         &auth.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch(tls.SubjectAltNames)},
+						ValidationContextSdsSecretConfig: authn_model.ConstructSdsSecretConfig(res.GetRootResourceName(), proxy),
+					},
+				}
+			}
+		}
+
+		if cb.httpProtocolOptions[c.Name] != nil && cb.httpProtocolOptions[c.Name].UpstreamProtocolOptions != nil {
+			// This is HTTP/2 cluster, advertise it with ALPN.
+			tlsContext.CommonTlsContext.AlpnProtocols = util.ALPNH2Only
+		}
+	}
+	return tlsContext, nil
+}
+
+func http2ProtocolOptions() *core.Http2ProtocolOptions {
+	return &core.Http2ProtocolOptions{
+		// Envoy default value of 100 is too low for data path.
+		MaxConcurrentStreams: &wrappers.UInt32Value{
+			Value: 1073741824,
+		},
+	}
+}
+
+// FIXME: there isn't a way to distinguish between unset values and zero values
+func (cb *ClusterBuilder) applyConnectionPool(mesh *meshconfig.MeshConfig, c *cluster.Cluster, settings *networking.ConnectionPoolSettings) {
+	if settings == nil {
+		return
+	}
+
+	threshold := getDefaultCircuitBreakerThresholds()
+	var idleTimeout *types.Duration
+
+	if settings.Http != nil {
+		if settings.Http.Http2MaxRequests > 0 {
+			// Envoy only applies MaxRequests in HTTP/2 clusters
+			threshold.MaxRequests = &wrappers.UInt32Value{Value: uint32(settings.Http.Http2MaxRequests)}
+		}
+		if settings.Http.Http1MaxPendingRequests > 0 {
+			// Envoy only applies MaxPendingRequests in HTTP/1.1 clusters
+			threshold.MaxPendingRequests = &wrappers.UInt32Value{Value: uint32(settings.Http.Http1MaxPendingRequests)}
+		}
+
+		if settings.Http.MaxRequestsPerConnection > 0 {
+			c.MaxRequestsPerConnection = &wrappers.UInt32Value{Value: uint32(settings.Http.MaxRequestsPerConnection)}
+		}
+
+		// FIXME: zero is a valid value if explicitly set, otherwise we want to use the default
+		if settings.Http.MaxRetries > 0 {
+			threshold.MaxRetries = &wrappers.UInt32Value{Value: uint32(settings.Http.MaxRetries)}
+		}
+
+		idleTimeout = settings.Http.IdleTimeout
+	}
+
+	if settings.Tcp != nil {
+		if settings.Tcp.ConnectTimeout != nil {
+			c.ConnectTimeout = gogo.DurationToProtoDuration(settings.Tcp.ConnectTimeout)
+		}
+
+		if settings.Tcp.MaxConnections > 0 {
+			threshold.MaxConnections = &wrappers.UInt32Value{Value: uint32(settings.Tcp.MaxConnections)}
+		}
+
+		applyTCPKeepalive(mesh, c, settings)
+	}
+
+	c.CircuitBreakers = &cluster.CircuitBreakers{
+		Thresholds: []*cluster.CircuitBreakers_Thresholds{threshold},
+	}
+
+	if idleTimeout != nil {
+		idleTimeoutDuration := gogo.DurationToProtoDuration(idleTimeout)
+		if cb.httpProtocolOptions[c.Name] == nil {
+			cb.httpProtocolOptions[c.Name] = &http.HttpProtocolOptions{}
+		}
+		commonOptions := cb.httpProtocolOptions[c.Name]
+		commonOptions.CommonHttpProtocolOptions = &core.HttpProtocolOptions{
+			IdleTimeout: idleTimeoutDuration,
+		}
+	}
 }
 
 // MergeTrafficPolicy returns the merged TrafficPolicy for a destination-level and subset-level policy on a given port.
@@ -198,7 +603,7 @@ func MergeTrafficPolicy(original, subsetPolicy *networking.TrafficPolicy, port *
 // buildDefaultCluster builds the default cluster and also applies default traffic policy.
 func (cb *ClusterBuilder) buildDefaultCluster(name string, discoveryType cluster.Cluster_DiscoveryType,
 	localityLbEndpoints []*endpoint.LocalityLbEndpoints, direction model.TrafficDirection,
-	port *model.Port, service *model.Service, allInstances []*model.ServiceInstance) *cluster.Cluster, buildClusterOpts {
+	port *model.Port, service *model.Service, allInstances []*model.ServiceInstance) *cluster.Cluster {
 	if allInstances == nil {
 		allInstances = cb.proxy.ServiceInstances
 	}
@@ -228,15 +633,15 @@ func (cb *ClusterBuilder) buildDefaultCluster(name string, discoveryType cluster
 	// For inbound clusters, the default traffic policy is used. For outbound clusters, the default traffic policy
 	// will be applied, which would be overridden by traffic policy specified in destination rule, if any.
 	opts := buildClusterOpts{
-		mesh:            cb.push.Mesh,
-		cluster:         c,
-		policy:          cb.defaultTrafficPolicy(discoveryType),
-		port:            port,
-		serviceAccounts: nil,
-		istioMtlsSni:    "",
-		clusterMode:     DefaultClusterMode,
-		direction:       direction,
-		proxy:           cb.proxy,
+		mesh:                cb.push.Mesh,
+		cluster:             c,
+		policy:              cb.defaultTrafficPolicy(discoveryType),
+		port:                port,
+		serviceAccounts:     nil,
+		istioMtlsSni:        "",
+		clusterMode:         DefaultClusterMode,
+		direction:           direction,
+		proxy:               cb.proxy,
 		httpProtocolOptions: nil,
 	}
 	// decides whether the cluster corresponds to a service external to mesh or not.
@@ -247,10 +652,10 @@ func (cb *ClusterBuilder) buildDefaultCluster(name string, discoveryType cluster
 		// otherwise, read this information from service object.
 		opts.meshExternal = service.MeshExternal
 	}
-	applyTrafficPolicy(opts)
+	cb.applyTrafficPolicy(opts)
 	addTelemetryMetadata(opts, service, direction, allInstances)
 	addNetworkingMetadata(opts, service, direction)
-	return c,opts
+	return c
 }
 
 // buildInboundClusterForPortOrUDS constructs a single inbound listener. The cluster will be bound to
@@ -281,7 +686,7 @@ func (cb *ClusterBuilder) buildInboundClusterForPortOrUDS(clusterPort int, bind 
 		localCluster.AltStatName = util.BuildStatPrefix(cb.push.Mesh.InboundClusterStatName,
 			string(instance.Service.Hostname), "", instance.ServicePort, instance.Service.Attributes)
 	}
-	setUpstreamProtocol(cb.proxy, localCluster, instance.ServicePort, model.TrafficDirectionInbound)
+	cb.setUpstreamProtocol(cb.proxy, localCluster, instance.ServicePort, model.TrafficDirectionInbound)
 
 	// When users specify circuit breakers, they need to be set on the receiver end
 	// (server side) as well as client side, so that the server has enough capacity
@@ -295,7 +700,7 @@ func (cb *ClusterBuilder) buildInboundClusterForPortOrUDS(clusterPort int, bind 
 			connectionPool, _, _, _ := selectTrafficPolicyComponents(MergeTrafficPolicy(nil, destinationRule.TrafficPolicy, instance.ServicePort))
 			// only connection pool settings make sense on the inbound path.
 			// upstream TLS settings/outlier detection/load balancer don't apply here.
-			applyConnectionPool(cb.push.Mesh, localCluster, connectionPool)
+			cb.applyConnectionPool(cb.push.Mesh, localCluster, connectionPool)
 			util.AddConfigInfoMetadata(localCluster.Metadata, cfg.Meta)
 		}
 	}
@@ -312,6 +717,53 @@ func (cb *ClusterBuilder) buildInboundClusterForPortOrUDS(clusterPort int, bind 
 		}
 	}
 	return localCluster
+}
+
+func (cb *ClusterBuilder) setUpstreamProtocol(node *model.Proxy, c *cluster.Cluster, port *model.Port, direction model.TrafficDirection) {
+	// Add use_downstream_protocol for sidecar proxy only if protocol sniffing is enabled.
+	// Since protocol detection is disabled for gateway and use_downstream_protocol is used
+	// under protocol detection for cluster to select upstream connection protocol when
+	// the service port is unnamed. use_downstream_protocol should be disabled for gateway.
+	if node.Type == model.SidecarProxy && ((util.IsProtocolSniffingEnabledForInboundPort(port) && direction == model.TrafficDirectionInbound) ||
+		(util.IsProtocolSniffingEnabledForOutboundPort(port) && direction == model.TrafficDirectionOutbound)) {
+		// Use downstream protocol. If the incoming traffic use HTTP 1.1, the
+		// upstream cluster will use HTTP 1.1, if incoming traffic use HTTP2,
+		// the upstream cluster will use HTTP2.
+		cb.setUseDownstreamProtocol(c)
+	} else if port.Protocol.IsHTTP2() {
+		cb.setH2Options(c)
+	}
+}
+
+// finalizeClusters does any final cluster field marshalling. This should be called
+// at the end before adding the cluster to list of clusters.
+func (cb *ClusterBuilder) finalizeClusters(clusters []*cluster.Cluster) []*cluster.Cluster {
+	// resolve cluster name conflicts. there can be duplicate cluster names if there are conflicting service definitions.
+	// for any clusters that share the same name the first cluster is kept and the others are discarded.
+	have := sets.Set{}
+	out := make([]*cluster.Cluster, 0, len(clusters))
+	for _, cluster := range clusters {
+		if !have.Contains(cluster.Name) {
+			out = append(out, cb.finalize(cluster))
+		} else {
+			cb.push.AddMetric(model.DuplicatedClusters, cluster.Name, cb.proxy.ID,
+				fmt.Sprintf("Duplicate cluster %s found while pushing CDS", cluster.Name))
+		}
+		have.Insert(cluster.Name)
+	}
+	return out
+}
+
+func (cb *ClusterBuilder) finalize(c *cluster.Cluster) *cluster.Cluster {
+	// Marshall Http Protocol options if they exist.
+	if httpOptions := cb.httpProtocolOptions[c.Name]; httpOptions != nil {
+		if moptions, err := ptypes.MarshalAny(httpOptions); err == nil {
+			c.TypedExtensionProtocolOptions = map[string]*any.Any{
+				v3.HttpProtocolOptionsType: moptions,
+			}
+		}
+	}
+	return c
 }
 
 func (cb *ClusterBuilder) buildLocalityLbEndpoints(proxyNetworkView map[string]bool, service *model.Service,
@@ -434,7 +886,7 @@ func (cb *ClusterBuilder) buildDefaultPassthroughCluster() *cluster.Cluster {
 		ProtocolSelection:    cluster.Cluster_USE_DOWNSTREAM_PROTOCOL,
 	}
 	passthroughSettings := &networking.ConnectionPoolSettings{}
-	applyConnectionPool(cb.push.Mesh, cluster, passthroughSettings)
+	cb.applyConnectionPool(cb.push.Mesh, cluster, passthroughSettings)
 	return cluster
 }
